Added Twitter timeline and header
I also need content to go on the left here. I was thinking either some
type of image branding (logo, decal) in the bottom left or a projects
list.

The image branding begs the question what image. I had the idea for a
circular drop spiral thing with the A in the middle. Purple and black
and white might be too much, maybe grey and slightly darker grey, or
maybe I throw the holi picture to kuler and see what happens.

The projects section would mean a sidebar navigation where you could
click through my projects. This would require the js to make the twitter
column disappear and display the project picture and description and
relevant links. I would definitely use flat or bootstrap-style buttons
for the links. It could also optionally make the about column disappear
so the projects section can use all that width. Not sure if I'd want to
do that though.